// typescript-node/v1/generated/model/openapiPaymentAccountGetOrCreateRequest.ts
/**
 * Resources
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { OpenapiPaymentAccountGetOrCreateRequestBankAccount } from './openapiPaymentAccountGetOrCreateRequestBankAccount';
import { OpenapiPaymentAccountGetOrCreateRequestHolder } from './openapiPaymentAccountGetOrCreateRequestHolder';
import { OpenapiPaymentAccountGetOrCreateRequestMobileMoney } from './openapiPaymentAccountGetOrCreateRequestMobileMoney';

export class OpenapiPaymentAccountGetOrCreateRequest {
    'bankAccount'?: OpenapiPaymentAccountGetOrCreateRequestBankAccount;
    /**
    * <span style=\"color:#e95f6a;\">required</span>  The ISO 3166 alpha-2 country code in which the payment account is registered.
    */
    'country'?: string;
    'holder'?: OpenapiPaymentAccountGetOrCreateRequestHolder;
    'mobileMoney'?: OpenapiPaymentAccountGetOrCreateRequestMobileMoney;
    /**
    * <span style=\"color:#e95f6a;\">required</span>  Either `MOBILE_MONEY` or `BANK_ACCOUNT`
    */
    'type'?: OpenapiPaymentAccountGetOrCreateRequest.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "bankAccount",
            "baseName": "bank_account",
            "type": "OpenapiPaymentAccountGetOrCreateRequestBankAccount"
        },
        {
            "name": "country",
            "baseName": "country",
            "type": "string"
        },
        {
            "name": "holder",
            "baseName": "holder",
            "type": "OpenapiPaymentAccountGetOrCreateRequestHolder"
        },
        {
            "name": "mobileMoney",
            "baseName": "mobile_money",
            "type": "OpenapiPaymentAccountGetOrCreateRequestMobileMoney"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "OpenapiPaymentAccountGetOrCreateRequest.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return OpenapiPaymentAccountGetOrCreateRequest.attributeTypeMap;
    }
}

export namespace OpenapiPaymentAccountGetOrCreateRequest {
    export enum TypeEnum {
        MobileMoney = <any> 'MOBILE_MONEY',
        BankAccount = <any> 'BANK_ACCOUNT'
    }
}