// typescript-node/v1/generated/model/openapiPaymentAccountGetOrCreateRequestHolder.ts
/**
 * Resources
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* <span style=\"color:#e95f6a;\">required</span>  The individual (or business) in whose name the bank account is registered
*/
export class OpenapiPaymentAccountGetOrCreateRequestHolder {
    /**
    * <span style=\"color:#e95f6a;\">required</span>  Either the INDIVIDUAL full name or the BUSINESS registered legal name
    */
    'name'?: string;
    /**
    * <span style=\"color:#e95f6a;\">required</span>  The holder legal entity type, whether an INDIVIDUAL person or a BUSINESS
    */
    'type'?: OpenapiPaymentAccountGetOrCreateRequestHolder.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "OpenapiPaymentAccountGetOrCreateRequestHolder.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return OpenapiPaymentAccountGetOrCreateRequestHolder.attributeTypeMap;
    }
}

export namespace OpenapiPaymentAccountGetOrCreateRequestHolder {
    export enum TypeEnum {
        Individual = <any> 'INDIVIDUAL',
        Business = <any> 'BUSINESS'
    }
}