// typescript-node/v1/generated/model/openapiPaymentAccountGetOrCreateResponseHolder.ts
/**
 * Resources
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class OpenapiPaymentAccountGetOrCreateResponseHolder {
    'name'?: string;
    'type'?: OpenapiPaymentAccountGetOrCreateResponseHolder.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "OpenapiPaymentAccountGetOrCreateResponseHolder.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return OpenapiPaymentAccountGetOrCreateResponseHolder.attributeTypeMap;
    }
}

export namespace OpenapiPaymentAccountGetOrCreateResponseHolder {
    export enum TypeEnum {
        Individual = <any> 'INDIVIDUAL',
        Business = <any> 'BUSINESS'
    }
}