// typescript-node/v1/generated/model/openapiPayoutCreateRequestAmount.ts
/**
 * Resources
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* <span style=\"color:#e95f6a;\">required</span>  The intended amount to send to either the payment_account_id or payment_account
*/
export class OpenapiPayoutCreateRequestAmount {
    /**
    * <span style=\"color:#e95f6a;\">required</span>  The currency associated with the amount value, formatted according to ISO 4217
    */
    'currency'?: string;
    /**
    * <span style=\"color:#e95f6a;\">required</span>
    */
    'value'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return OpenapiPayoutCreateRequestAmount.attributeTypeMap;
    }
}