// typescript-node/v1/generated/model/openapiPayoutCreateRequestSender.ts
/**
 * Resources
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { OpenapiPayoutCreateRequestSenderAddress } from './openapiPayoutCreateRequestSenderAddress';
import { OpenapiPayoutCreateRequestSenderIndividual } from './openapiPayoutCreateRequestSenderIndividual';

/**
* An optional property for incorporating sender details. Please refer to the \"💡 **Compliance requirements**\" documentation above.
*/
export class OpenapiPayoutCreateRequestSender {
    'address'?: OpenapiPayoutCreateRequestSenderAddress;
    'individual'?: OpenapiPayoutCreateRequestSenderIndividual;
    /**
    * <span style=\"color:#e95f6a;\">required if sender is provided</span>  The individual full name or registered business legal name
    */
    'name'?: string;
    /**
    * <span style=\"color:#e95f6a;\">required if sender is provided</span>  Either `INDIVIDUAL` or `BUSINESS`
    */
    'type'?: OpenapiPayoutCreateRequestSender.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "address",
            "baseName": "address",
            "type": "OpenapiPayoutCreateRequestSenderAddress"
        },
        {
            "name": "individual",
            "baseName": "individual",
            "type": "OpenapiPayoutCreateRequestSenderIndividual"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "OpenapiPayoutCreateRequestSender.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return OpenapiPayoutCreateRequestSender.attributeTypeMap;
    }
}

export namespace OpenapiPayoutCreateRequestSender {
    export enum TypeEnum {
        Individual = <any> 'INDIVIDUAL',
        Business = <any> 'BUSINESS'
    }
}